// shared/addresses/types.ts
export type Gender = "f" | "m" | "n";

export interface ProtoWordConfig {
  normalizedValue?: string;
  beautifiedValue?: string;
  aliases?: Readonly<string[]>;
}

export interface WordConfig extends ProtoWordConfig {
  normalizedValue: string;
  wordReplacements?: Readonly<string[]>;
}

export type Designation =
  | "region"
  | "county"
  | "settlement"
  | "place"
  | "district"
  | "street"
  | "house"
  | "housePart";

export interface DesignationConfig extends WordConfig {
  designation: Designation;
  gender: Gender;
  canBePartOfName?: boolean;
  canBeSkippedIfAloneInSection?: boolean;
  alwaysGoesBeforeName?: boolean;
}
// ...
export interface AddressWordBase {
  nodeType: "word";
  value: string;
}

export interface AddressNodeWithUnclassifiedWord extends AddressWordBase {
  wordType: "unclassified";
}

export interface AddressNodeWithInitial extends AddressWordBase {
  wordType: "initial";
}

export interface AddressNodeWithDesignation extends AddressWordBase {
  wordType: "designation";
}
// ...
export interface WordReplacementConfig {
  detached?: boolean;
  from: string | string[];
  to: string;
  silenceNormalizationWarning?: boolean;
}